Separate routes from the app initialization
Notice that in `main.py` i imported the routes after initialize the flask app *to avoid circular import*
